backend/requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.0.3
PyJWT==2.8.0
Werkzeug==2.2.3
passlib[bcrypt]==1.7.4